clippy and make benchnmark crate usable as a dependency
