+++ benchmarks/Cargo.toml
@@ -24,7 +24,6 @@ authors = ["Apache Arrow <dev@arrow.apache.org>"]
 homepage = "https://github.com/apache/arrow-datafusion"
 repository = "https://github.com/apache/arrow-datafusion"
 license = "Apache-2.0"
-publish = false
 rust-version = "1.62"
 
 [features]
